refactor controllers namespace
