+++ app/controllers/spid-rails/application_controller.rb
@@ -1,0 +1,9 @@
+module Spid
+  module Rails
+
+    class ApplicationController < ActionController::Base
+      protect_from_forgery with: :exception
+    end
+    
+  end
+end
