+++ app/controllers/spid-rails/metadata_controller.rb
@@ -1,0 +1,17 @@
+require_dependency "spid-rails/application_controller"
+
+# Metadata del Service Provider
+module Spid
+  module Rails
+
+    class MetadataController < ApplicationController
+
+      def show
+        metadata = SpidRails::Metadata.create(host: main_app.root_url)
+        render xml: metadata.to_xml
+      end
+      
+    end
+
+  end
+end
